Extract user scopes tests to a separate file
